Implement Lexer Functionality
Lexer is responsible for taking a raw text input and transferring it
into a vector of strings. Current all white space is ignore, all words
are accepted, only ; , ( ) . " ' are recognized punctuation.
Parser takes the lexer input and in the future will handle the
interaction between the DBMS and the input code.