Implement user input handling
User can now input queries into the DBMS. This also handles cleaning the
input for what the lexer and parser expects. This function is the only
way for the user to interact with the database so this also handles the
user quitting if they desire.
Get_query() does allow users to use tabes and multiple lines for visual
reasons but after input these are cleaned out of the text.
As of now there is no functionality associated with user input yet.