Fix bug where a ; database would be created
This bug occured when a user would forget to input database name. I
previously checked to see if it was an end token to bypass this case but
semicolon will always come before end token. In this case it would see a
semicolon and move on to create a database with ; as the identifier.
Solution was to just change the check to be ; instead.
Additionally fixed the create table error text to correctly say expected
table instead of database.