Implement show database statement and bugfix
Implementation of show database was overall simple but in the process of
testing this new addition encountered a critical bug with how we were
assigning new parser objects.
Previously we were shallow moving parser objects. This behavior was
hidden previously because as long as the current query matched the
syntax structure of the first query the program runs fine.
Solution was to implement a copy constructor that copys the token vec
but does not copy the iterator otherwise the current object would be
pointing to a soon to be deleted pointer. Instead assign the iterator to
the freshly assigned token vec.