Implment looking ahead for parser and bugfix
Now give the ability to look ahead in the parser for commands that
require knowing what the next token is before executing. The secondary
purpose for this is to create an abstraction in our current
implementation to prevent redundant code.
Also have a lexer bugfix in this commit. There was a boundary situation
where the lexer has a word ready but the string is done. There is no
call to add token in this situation so the token never gets added before
the end of the process. Fixed it by simply having an add token call
after the loop.