Implement insert statement
Insert statement handles grabbing user input that is to be inserted into
a table. This process involed updated table and database as the DBMS
hands a string vector to database but database has no business making
records so it has to pass this down to table and then table takes it
from there.