Bplustree object deletion and sorting bugfix
Before I implemented nodes without heap allocation but this proved to be
an issue for object deletion as nodes are passed around and so
situations nodes could get deleted without handling their associated
pointers. Anyway the change I made here is node_block now stores node
pointers. Nodes are now created using new and the pointer is set to a
node_block. This allowed a clean recursive object deletion method.

Through this process a bug was discovered with sorting. As we are now
passing pointers to the sorting algorithm it is not longer sorting based
on the data field but now the actual value of the pointer which is
incorrect. To fix this we created a new sorting operator that handles
accessing the data through the pointer.