Implement Simple Select
This commit is a bit irresponsible on my part. I was focued on adding
one thing but this addition involved alot of moving parts.

General Idea is am now properly implementing the simple select
statement. This requires creating an entire new relation(Table) but i
did not have code to allow for this. The changes involved here are two.

The actually syntax checking of the simple select statement which was
rather simple.

The second being building the infrastructure to allow for a table to
make a copy of all its records and place it into a vector that can be
passed on to a new table. This was an intensive process that involved
every key componenet of the DBMS to have new functionality.